wip: split those new modules into common and metrologicals and introduce utils and network as modules
