+++ agentMET4FOF/agents/base_agents.py
@@ -1,42 +1,25 @@
 import base64
-import copy
-import csv
 import datetime
-import re
-import sys
 import time
 from collections import deque
 from io import BytesIO
-from threading import Timer
-from typing import Any, Callable, Dict, Iterable, List, Optional, Sized, Tuple, Union
+from typing import Any, Callable, Dict, List, Optional, Union
 
 import matplotlib.figure
 import matplotlib.pyplot as plt
 import mpld3
-import networkx as nx
 import numpy as np
-import pandas as pd
-from mesa import Agent as MesaAgent, Model
-from mesa.time import BaseScheduler
-from osbrain import Agent as osBrainAgent, NSProxy, run_agent, run_nameserver
-from pandas import DataFrame
-from plotly import graph_objs as go, tools as tls
+from mesa import Agent as MesaAgent
+from osbrain import Agent as osBrainAgent
+from plotly import tools as tls
 from plotly.graph_objs import Scatter
-from time_series_buffer import TimeSeriesBuffer
-from time_series_metadata.scheme import MetaData
 
-from ..dashboard.default_network_stylesheet import default_agent_network_stylesheet
+from .utils import AgentBuffer
 from ..streams.base_streams import DataStreamMET4FOF
-
 
 __all__ = [
     "AgentMET4FOF",
     "MonitorAgent",
-    "AgentNetwork",
-    "AgentBuffer",
-    "MetrologicalAgent",
-    "MetrologicalAgentBuffer",
-    "MetrologicalMonitorAgent",
 ]
 
 
@@ -899,1200 +882,6 @@ class AgentMET4FOF(MesaAgent, osBrainAgent):
             del self
 
 
-class AgentBuffer:
-    """Buffer class which is instantiated in every agent to store data incrementally
-
-    This buffer is necessary to handle multiple inputs coming from agents.
-
-    We can access the buffer like a dict with exposed functions such as .values(),
-    .keys() and .items(). The actual dict object is stored in the variable
-    :attr:`buffer`.
-
-    Attributes
-    ----------
-    buffer : dict of iterables or dict of dicts of iterables
-        The buffer can be a dict of iterables, or a dict of dict of iterables for nested
-        named data. The keys are the names of agents.
-    buffer_size :
-        The total number of elements to be stored in the agent :attr:`buffer`
-    supported_datatypes : list of types
-        List of all types supported and thus properly handled by the buffer. Defaults to
-        :class:`np.ndarray <NumPy:numpy.ndarray>`, list and Pandas
-        :class:`DataFrame <Pandas:pandas.DataFrame>`
-    """
-
-    def __init__(self, buffer_size: int = 1000):
-        """Initialise a new agent buffer object
-
-        Parameters
-        ----------
-        buffer_size: int
-            Length of buffer allowed.
-        """
-        self.buffer = {}
-        self.buffer_size = buffer_size
-        self.supported_datatypes = [list, pd.DataFrame, np.ndarray]
-
-    def __getitem__(self, key):
-        return self.buffer[key]
-
-    def check_supported_datatype(self, obj: object) -> bool:
-        """Checks whether `value` is an object of one of the supported data types
-
-        Parameters
-        ----------
-        obj : object
-            Value to be checked
-
-        Returns
-        ------
-        result : boolean
-            True if value is an object of one of the supported data types, False if not
-        """
-        for supported_datatype in self.supported_datatypes:
-            if isinstance(obj, supported_datatype):
-                return True
-        return False
-
-    def update(
-        self,
-        agent_from: Union[Dict[str, Union[np.ndarray, list, pd.DataFrame]], str],
-        data: Union[np.ndarray, list, pd.DataFrame, float, int] = None,
-    ):
-        """Overrides data in the buffer dict keyed by ``agent_from`` with value ``data``
-
-        If ``data`` is a single value, this converts it into a list first before storing
-        in the buffer dict.
-
-        Parameters
-        ----------
-        agent_from : str
-            Name of agent sender
-        data : np.ndarray, DataFrame, list, float or int
-            New incoming data
-        """
-        # handle if data type nested in dict
-        if isinstance(data, dict):
-            # check for each value datatype
-            for key, value in data.items():
-                # if the value is not list types, turn it into a list of single value
-                # i.e [value]
-                if not self.check_supported_datatype(value):
-                    data[key] = [value]
-        elif not self.check_supported_datatype(data):
-            data = [data]
-        self.buffer.update({agent_from: data})
-        return self.buffer
-
-    def _concatenate(
-        self,
-        iterable: Union[np.ndarray, list, pd.DataFrame],
-        data: Union[np.ndarray, list, DataFrame],
-        concat_axis: int = 0,
-    ) -> Iterable:
-        """Concatenate the given ``iterable`` with ``data``
-
-        Handles the concatenation function depending on the datatype, and truncates it
-        if the buffer is filled to `buffer_size`.
-
-        Parameters
-        ----------
-        iterable : any in supported_datatype
-            The current buffer to be concatenated with.
-
-        data : np.ndarray, DataFrame, list
-            New incoming data
-
-        Returns
-        any in supported_datatype
-            the original buffer with the data appended
-        """
-        # handle list
-        if isinstance(iterable, list):
-            iterable += data
-            # check if exceed memory buffer size, remove the first n elements which
-            # exceeded the size
-            if len(iterable) > self.buffer_size:
-                truncated_element_index = len(iterable) - self.buffer_size
-                iterable = iterable[truncated_element_index:]
-
-        # handle if data type is np.ndarray
-        elif isinstance(iterable, np.ndarray):
-            iterable = np.concatenate((iterable, data), axis=concat_axis)
-            if len(iterable) > self.buffer_size:
-                truncated_element_index = len(iterable) - self.buffer_size
-                iterable = iterable[truncated_element_index:]
-
-        # handle if data type is pd.DataFrame
-        elif isinstance(iterable, pd.DataFrame):
-            iterable = pd.concat([iterable, data], ignore_index=True, axis=concat_axis)
-            if len(iterable) > self.buffer_size:
-                truncated_element_index = len(iterable) - self.buffer_size
-                iterable = iterable.truncate(before=truncated_element_index)
-        return iterable
-
-    def buffer_filled(self, agent_from: Optional[str] = None) -> bool:
-        """Checks whether buffer is filled, by comparing against the :attr:`buffer_size`
-
-        For nested dict, this returns True if any of the iterables is beyond the
-        :attr:`buffer_size`. For any of the dict values , which is not one of
-        :attr:`supported_datatypes` this returns None.
-
-        Parameters
-        ----------
-        agent_from : str, optional
-            Name of input agent in the buffer dict to be looked up. If ``agent_from``
-            is not provided, we check for all iterables in the buffer (default).
-
-        Returns
-        -------
-        bool or None
-            True if either the or any of the iterables has reached
-            :attr:`buffer_size` or None in case none of the values is of one of the
-            supported datatypes. False if all present iterable can take at least
-            one more element according to :attr:`buffer_size`.
-        """
-        if agent_from is None:
-            return any([self._iterable_filled(iterable) for iterable in self.values()])
-        elif isinstance(self[agent_from], dict):
-            return any(
-                [
-                    self._iterable_filled(iterable)
-                    for iterable in self[agent_from].values()
-                ]
-            )
-        else:
-            return self._iterable_filled(self[agent_from])
-
-    def _iterable_filled(self, iterable: Sized) -> Union[bool, None]:
-        """Internal method for checking on length of iterables of supported types
-
-        Parameters
-        ----------
-        iterable : Any
-            Expected to be an iterable of one of the supported datatypes but could be
-            any.
-
-        Returns
-        -------
-        bool or None
-            True if the iterable is of one of the supported datatypes and has reached
-            :attr:`buffer_size` in length or False if not or None in case it is not of
-            one of the supported datatypes.
-        """
-        if self.check_supported_datatype(iterable):
-            if len(iterable) >= self.buffer_size:
-                return True
-            return False
-
-    def popleft(self, n: Optional[int] = 1) -> Union[Dict, np.ndarray, list, DataFrame]:
-        """Pops the first n entries in the buffer
-
-        Parameters
-        ---------
-        n : int
-            Number of elements to retrieve from buffer
-
-        Returns
-        -------
-        dict, :class:`np.ndarray <NumPy:numpy.ndarray>`, list or Pandas
-        :class:`DataFrame <Pandas:pandas.DataFrame>`
-            The retrieved elements
-        """
-        popped_buffer = copy.copy(self.buffer)
-        remaining_buffer = copy.copy(self.buffer)
-        if isinstance(popped_buffer, dict):
-            for key, value in popped_buffer.items():
-                value, remaining_buffer[key] = self._popleft(value, n)
-        else:
-            popped_buffer, remaining_buffer = self._popleft(popped_buffer, n)
-        self.buffer = remaining_buffer
-        return popped_buffer
-
-    def _popleft(
-        self, iterable: Union[np.ndarray, list, DataFrame], n: Optional[int] = 1
-    ) -> Tuple[Union[np.ndarray, list, DataFrame], Union[np.ndarray, list, DataFrame]]:
-        """Internal handler of the actual popping mechanism based on type of iterable
-
-        Parameters
-        ---------
-        n : int
-            Number of elements to retrieve from buffer.
-        iterable : any in :attr:`supported_datatypes`
-            The current buffer to retrieve from.
-
-        Returns
-        -------
-        2-tuple of each either one of :class:`np.ndarray <NumPy:numpy.ndarray>`,
-        list or Pandas :class:`DataFrame <Pandas:pandas.DataFrame>`
-            The retrieved elements and the residual items in the buffer
-        """
-        popped_item = 0
-        if isinstance(iterable, list):
-            popped_item = iterable[:n]
-            iterable = iterable[n:]
-        elif isinstance(iterable, np.ndarray):
-            popped_item = iterable[:n]
-            iterable = iterable[n:]
-        elif isinstance(iterable, pd.DataFrame):
-            popped_item = iterable.iloc[:n]
-            iterable = iterable.iloc[n:]
-        return popped_item, iterable
-
-    def clear(self, agent_from: Optional[str] = None):
-        """Clears the data in the buffer
-
-        Parameters
-        ----------
-        agent_from : str, optional
-            Name of agent, if ``agent_from`` is not given, the entire buffer is
-            flushed. (default)
-        """
-        if agent_from is None:
-            self.buffer = {}
-        elif agent_from in self.buffer:
-            del self.buffer[agent_from]
-
-    def store(
-        self,
-        agent_from: Union[Dict[str, Union[np.ndarray, list, pd.DataFrame]], str],
-        data: Union[np.ndarray, list, pd.DataFrame, float, int] = None,
-        concat_axis: Optional[int] = 0,
-    ):
-        """Stores data into :attr:`buffer` with the received message
-
-        Checks if sender agent has sent any message before. If it did, then append,
-        otherwise create new entry for it.
-
-        Parameters
-        ----------
-        agent_from : dict | str
-            if type is dict, we expect it to be the agentMET4FOF dict message to be
-            compliant with older code (keys ``from`` and ``data`` present'), otherwise
-            we expect it to be name of agent sender and ``data`` will need to be passed
-            as parameter
-        data : np.ndarray, DataFrame, list, float or int
-            Not used if ``agent_from`` is a dict. Otherwise ``data`` is compulsory.
-        concat_axis : int, optional
-            axis to concatenate on with the buffering for numpy arrays.
-            Default is 0.
-        """
-        # Store into a separate variables, it will be used frequently later for the
-        # type checks. If first argument is the agentMET4FOF dict message in old format
-        if isinstance(agent_from, dict):
-            message_from = agent_from["from"]
-            message_data = agent_from["data"]
-        # ... otherwise, we expect the name of agent_sender and the data to be passed.
-        else:
-            message_from = agent_from
-            message_data = data
-
-        # check if sender agent has sent any message before:
-        # if it did,then append, otherwise create new entry for the input agent
-        if message_from not in self.buffer:
-            self.update(message_from, message_data)
-            return 0
-
-        # otherwise 'sender' exists in memory, handle appending
-        # acceptable data types : list, dict, ndarray, dataframe, single values
-
-        # handle nested data in dict
-        if isinstance(message_data, dict):
-            for key, value in message_data.items():
-                # if it is a single value, then we convert it into a single element list
-                if not self.check_supported_datatype(value):
-                    value = [value]
-                # check if the key exist
-                # if it does, then append
-                if key in self.buffer[agent_from].keys():
-                    self.buffer[agent_from][key] = self._concatenate(
-                        self.buffer[agent_from][key], value, concat_axis
-                    )
-                # otherwise, create new entry
-                else:
-                    self.buffer[agent_from].update({key: value})
-        else:
-            if not self.check_supported_datatype(message_data):
-                message_data = [message_data]
-            self.buffer[agent_from] = self._concatenate(
-                self.buffer[agent_from], message_data, concat_axis
-            )
-
-    def values(self):
-        """Interface to access the internal dict's values()"""
-        return self.buffer.values()
-
-    def items(self):
-        """Interface to access the internal dict's items()"""
-        return self.buffer.items()
-
-    def keys(self):
-        """Interface to access the internal dict's keys()"""
-        return self.buffer.keys()
-
-
-class _AgentController(AgentMET4FOF):
-    """Unique internal agent to provide control to other agents
-
-    Automatically instantiated when starting server. Provides global control to all
-    agents in network.
-    """
-
-    def init_parameters(self, ns=None, backend="osbrain", mesa_model="", log_mode=True):
-        self.backend = backend
-        self.states = {0: "Idle", 1: "Running", 2: "Pause", 3: "Stop"}
-        self.current_state = "Idle"
-        self.ns = ns
-        self.G = nx.DiGraph()
-        self._logger = None
-        self.coalitions = []
-        self.log_mode = log_mode
-
-        if backend == "mesa":
-            self.mesa_model = mesa_model
-
-    def start_mesa_timer(self, mesa_update_interval):
-        class RepeatTimer:
-            def __init__(self, t, repeat_function):
-                self.t = t
-                self.repeat_function = repeat_function
-                self.thread = Timer(self.t, self.handle_function)
-
-            def handle_function(self):
-                self.repeat_function()
-                self.thread = Timer(self.t, self.handle_function)
-                self.thread.start()
-
-            def start(self):
-                self.thread.start()
-
-            def cancel(self):
-                self.thread.cancel()
-
-        self.mesa_update_interval = mesa_update_interval
-        self.mesa_timer = RepeatTimer(
-            t=mesa_update_interval, repeat_function=self.mesa_model.step
-        )
-        self.mesa_timer.start()
-
-    def stop_mesa_timer(self):
-        if self.mesa_timer:
-            self.mesa_timer.cancel()
-            del self.mesa_timer
-
-    def step_mesa_model(self):
-        self.mesa_model.step()
-
-    def get_mesa_model(self):
-        return self.mesa_model
-
-    def get_agent(self, agentName=""):
-        if self.backend == "osbrain":
-            return self.ns.proxy(agentName)
-        elif self.backend == "mesa":
-            return self.mesa_model.get_agent(agentName)
-
-    def get_agentType_count(self, agentType):
-        num_count = 1
-        agentType_name = str(agentType.__name__)
-        agent_names = self.agents()
-        if len(agent_names) != 0:
-            for agentName in agent_names:
-                current_agent_type = self.get_agent(agentName).get_attr("AgentType")
-                if current_agent_type == agentType_name:
-                    num_count += 1
-        return num_count
-
-    def get_agent_name_count(self, new_agent_name):
-        num_count = 1
-        agent_names = self.agents()
-        if len(agent_names) != 0:
-            for agentName in agent_names:
-                if new_agent_name in agentName:
-                    num_count += 1
-        return num_count
-
-    def generate_module_name_byType(self, agentType):
-        # handle agent type
-        if isinstance(agentType, str):
-            name = agentType
-        else:
-            name = agentType.__name__
-        name += "_" + str(self.get_agentType_count(agentType))
-        return name
-
-    def generate_module_name_byUnique(self, agent_name):
-        name = agent_name
-        agent_copy_count = self.get_agent_name_count(
-            agent_name
-        )  # number of agents with same name
-        if agent_copy_count > 1:
-            name += "(" + str(self.get_agent_name_count(agent_name)) + ")"
-        return name
-
-    def add_agent(
-        self,
-        name=" ",
-        agentType=AgentMET4FOF,
-        log_mode=True,
-        buffer_size=1000,
-        ip_addr=None,
-        loop_wait=None,
-        **kwargs,
-    ):
-        try:
-            if ip_addr is None:
-                ip_addr = "localhost"
-
-            if name == " ":
-                new_name = self.generate_module_name_byType(agentType)
-            else:
-                new_name = self.generate_module_name_byUnique(name)
-
-            # actual instantiation of agent, depending on backend
-            if self.backend == "osbrain":
-                new_agent = self._add_osbrain_agent(
-                    name=new_name,
-                    agentType=agentType,
-                    log_mode=log_mode,
-                    buffer_size=buffer_size,
-                    ip_addr=ip_addr,
-                    loop_wait=loop_wait,
-                    **kwargs,
-                )
-            elif self.backend == "mesa":
-                # handle osbrain and mesa here
-                new_agent = self._add_mesa_agent(
-                    name=new_name,
-                    agentType=agentType,
-                    buffer_size=buffer_size,
-                    log_mode=log_mode,
-                    **kwargs,
-                )
-            return new_agent
-        except Exception as e:
-            self.log_info("ERROR:" + str(e))
-
-    def _add_osbrain_agent(
-        self,
-        name=" ",
-        agentType=AgentMET4FOF,
-        log_mode=True,
-        buffer_size=1000,
-        ip_addr=None,
-        loop_wait=None,
-        **kwargs,
-    ):
-        new_agent = run_agent(
-            name,
-            base=agentType,
-            attributes=dict(log_mode=log_mode, buffer_size=buffer_size),
-            nsaddr=self.ns.addr(),
-            addr=ip_addr,
-        )
-        new_agent.init_parameters(**kwargs)
-        new_agent.init_agent(buffer_size=buffer_size, log_mode=log_mode)
-        new_agent.init_agent_loop(loop_wait)
-        if log_mode:
-            new_agent.set_logger(self._get_logger())
-        return new_agent
-
-    def _add_mesa_agent(
-        self,
-        name=" ",
-        agentType=AgentMET4FOF,
-        log_mode=True,
-        buffer_size=1000,
-        **kwargs,
-    ):
-        new_agent = agentType(
-            name=name, backend=self.backend, mesa_model=self.mesa_model
-        )
-        new_agent.init_parameters(**kwargs)
-        new_agent.init_agent(buffer_size=buffer_size, log_mode=log_mode)
-        new_agent = self.mesa_model.add_agent(new_agent)
-        return new_agent
-
-    def get_agents_stylesheets(self, agent_names):
-        # for customising display purposes in dashboard
-        agents_stylesheets = []
-        for agent in agent_names:
-            try:
-                stylesheet = self.get_agent(agent).get_attr("stylesheet")
-                agents_stylesheets.append({"stylesheet": stylesheet})
-            except Exception as e:
-                self.log_info("Error:" + str(e))
-        return agents_stylesheets
-
-    def agents(self, exclude_names=["AgentController", "Logger"]):
-        if self.backend == "osbrain":
-            agent_names = [
-                name for name in self.ns.agents() if name not in exclude_names
-            ]
-        else:
-            agent_names = self.mesa_model.agents()
-        return agent_names
-
-    def update_networkx(self):
-        agent_names = self.agents()
-        edges = self.get_latest_edges(agent_names)
-
-        if (
-            len(agent_names) != self.G.number_of_nodes()
-            or len(edges) != self.G.number_of_edges()
-        ):
-            agent_stylesheets = self.get_agents_stylesheets(agent_names)
-            new_G = nx.DiGraph()
-            new_G.add_nodes_from(list(zip(agent_names, agent_stylesheets)))
-            new_G.add_edges_from(edges)
-            self.G = new_G
-
-    def get_networkx(self):
-        return self.G
-
-    def get_latest_edges(self, agent_names):
-        edges = []
-        for agent_name in agent_names:
-            temp_agent = self.get_agent(agent_name)
-            output_agent_channels = temp_agent.get_attr("Outputs_agent_channels")
-            temp_output_agents = list(output_agent_channels.keys())
-            temp_output_channels = list(output_agent_channels.values())
-
-            for output_agent_name, output_agent_channel in zip(
-                temp_output_agents, temp_output_channels
-            ):
-                edges += [
-                    (
-                        agent_name,
-                        output_agent_name,
-                        {"channel": str(output_agent_channel)},
-                    )
-                ]
-        return edges
-
-    def _get_logger(self):
-        """
-        Internal method to access the Logger relative to the nameserver
-        """
-        if self._logger is None:
-            self._logger = self.ns.proxy("Logger")
-        return self._logger
-
-    def add_coalition(self, new_coalition):
-        """
-        Instantiates a coalition of agents.
-        """
-        self.coalitions.append(new_coalition)
-        return new_coalition
-
-    def del_coalition(self):
-        self.coalitions = []
-
-    def add_coalition_agent(self, name, agents=[]):
-        """
-        Add agents into the coalition
-        """
-        # update coalition
-        for coalition_i, coalition in enumerate(self.coalitions):
-            if coalition.name == name:
-                for agent in agents:
-                    self.coalitions[coalition_i].add_agent(agent)
-
-    def remove_coalition_agent(self, coalition_name, agent_name=""):
-        """
-        Remove agent from coalition
-        """
-        # update coalition
-        for coalition_i, coalition in enumerate(self.coalitions):
-            if coalition.name == coalition_name:
-                self.coalitions[coalition_i].remove_agent(agent_name)
-
-    def get_coalition(self, name):
-        """
-        Gets the coalition based on provided name
-        """
-        for coalition_i, coalition in enumerate(self.coalitions):
-            if coalition.name == name:
-                return coalition
-        return -1
-
-
-class MesaModel(Model):
-    """A MESA Model"""
-
-    def __init__(self):
-        self.schedule = BaseScheduler(self)
-
-    def add_agent(self, agent: MesaAgent):
-        self.schedule.add(agent)
-        return agent
-
-    def get_agent(self, agentName: str):
-        agent = next((x for x in self.schedule.agents if x.name == agentName), None)
-        return agent
-
-    def step(self):
-        """Advance the model by one step."""
-        self.schedule.step()
-
-    def agents(self):
-        return [agent.name for agent in self.schedule.agents]
-
-    def shutdown(self):
-        """Shutdown entire MESA model with all agents and schedulers"""
-        for agent in self.agents():
-            agent_obj = self.get_agent(agent)
-            agent_obj.shutdown()
-
-
-class AgentNetwork:
-    """Object for starting a new Agent Network or connect to an existing Agent Network
-
-    An existing Agent Network can be specified by ip & port. Provides function to
-    add agents, (un)bind agents, query agent network state, set global agent states
-    Interfaces with an internal _AgentController which is hidden from user.
-    """
-
-    def __init__(
-        self,
-        ip_addr="127.0.0.1",
-        port=3333,
-        connect=False,
-        log_filename="log_file.csv",
-        dashboard_modules=True,
-        dashboard_extensions=[],
-        dashboard_update_interval=3,
-        dashboard_max_monitors=10,
-        dashboard_port=8050,
-        backend="osbrain",
-        mesa_update_interval=0.1,
-        network_stylesheet=default_agent_network_stylesheet,
-        **dashboard_kwargs,
-    ):
-        """
-        Parameters
-        ----------
-        ip_addr: str
-            Ip address of server to connect/start
-        port: int
-            Port of server to connect/start
-        connect: bool
-            False sets Agent network to connect mode and will connect to specified
-            address, True (Default) sets Agent network to initially try to connect
-            and if it cant find one, it will start a new server at specified address
-        log_filename: str
-            Name of log file, acceptable csv format. It will be saved locally,
-            in the same folder as the python script in which this AgentNetwork is
-            instantiated on. If set to None or False, then will not save in a file.
-            Note that the overhead of updating the log file can be huge, especially
-            for high number of agents and large data transmission.
-        dashboard_modules : list of modules , modules or bool
-            Accepts list of modules which contains the AgentMET4FOF and
-            DataStreamMET4FOF derived classes. If set to True, will initiate the
-            dashboard with default agents in AgentMET4FOF
-        dashboard_update_interval : int
-            Regular interval (seconds) to update the dashboard graphs
-        dashboard_max_monitors : int
-            Due to complexity in managing and instantiating dynamic figures,
-            a maximum number of monitors is specified first and only the each
-            Monitor Agent will occupy one of these figures.
-        dashboard_port: int
-            Port of the dashboard to be hosted on. By default is port 8050.
-        **dashboard_kwargs
-            Additional key words to be passed in initialising the dashboard
-        """
-
-        self.backend = backend
-        self.ip_addr = ip_addr
-        self.port = port
-        self._controller = None
-        self._logger = None
-        self.log_filename = log_filename
-
-        self.mesa_update_interval = mesa_update_interval
-        if connect:
-            self.is_parent_mesa = False
-        else:
-            self.is_parent_mesa = True
-
-        if type(self.log_filename) == str and ".csv" in self.log_filename:
-            self.save_logfile = True
-        else:
-            self.save_logfile = False
-
-        # handle different choices of backends
-        if self.backend == "osbrain":
-            if connect:
-                self.connect(ip_addr, port, verbose=False)
-            else:
-                self.connect(ip_addr, port, verbose=False)
-                if self.ns == 0:
-                    self.start_server_osbrain(ip_addr, port)
-        elif self.backend == "mesa":
-            self.start_server_mesa()
-        else:
-            raise NotImplementedError(
-                "Backend has not been implemented. Valid choices are 'osbrain' and "
-                "'mesa'."
-            )
-
-        if isinstance(dashboard_extensions, list) == False:
-            dashboard_extensions = [dashboard_extensions]
-
-        # handle instantiating the dashboard
-        # if dashboard_modules is False, the dashboard will not be launched
-        if dashboard_modules is not False:
-            from ..dashboard.Dashboard_agt_net import Dashboard_agt_net
-
-            # Initialize common dashboard parameters for both types of dashboards
-            # corresponding to different backends.
-            dashboard_params = {
-                "dashboard_modules": dashboard_modules,
-                "dashboard_layouts": [Dashboard_agt_net] + dashboard_extensions,
-                "dashboard_update_interval": dashboard_update_interval,
-                "max_monitors": dashboard_max_monitors,
-                "ip_addr": ip_addr,
-                "port": dashboard_port,
-                "agentNetwork": self,
-                "network_stylesheet": network_stylesheet,
-            }
-            dashboard_params.update(dashboard_kwargs)
-
-            # Initialize dashboard process/thread.
-            if self.backend == "osbrain":
-                from ..dashboard.Dashboard import AgentDashboardThread
-
-                self.dashboard_proc = AgentDashboardThread(**dashboard_params)
-            elif self.backend == "mesa":
-                from ..dashboard.Dashboard import AgentDashboardThread
-
-                self.dashboard_proc = AgentDashboardThread(**dashboard_params)
-            self.dashboard_proc.start()
-        else:
-            self.dashboard_proc = None
-
-    def connect(self, ip_addr="127.0.0.1", port=3333, verbose=True):
-        """
-        Only for osbrain backend. Connects to an existing AgentNetwork.
-
-        Parameters
-        ----------
-        ip_addr: str
-            IP Address of server to connect to
-
-        port: int
-            Port of server to connect to
-        """
-        try:
-            self.ns = NSProxy(nsaddr=ip_addr + ":" + str(port))
-        except:
-            if verbose:
-                print("Unable to connect to existing NameServer...")
-            self.ns = 0
-
-    def start_server_osbrain(self, ip_addr: str = "127.0.0.1", port: int = 3333):
-        """Starts a new AgentNetwork for osBrain and initializes :attr:`_controller`
-
-        Parameters
-        ----------
-        ip_addr: str
-            IP Address of server to start
-
-        port: int
-            Port of server to start
-        """
-
-        print("Starting NameServer...")
-        self.ns = run_nameserver(addr=ip_addr + ":" + str(port))
-        if len(self.ns.agents()) != 0:
-            self.ns.shutdown()
-            self.ns = run_nameserver(addr=ip_addr + ":" + str(port))
-        self._controller = run_agent(
-            "AgentController",
-            base=_AgentController,
-            attributes=dict(log_mode=True),
-            nsaddr=self.ns.addr(),
-            addr=ip_addr,
-        )
-        self._logger = run_agent("Logger", base=_Logger, nsaddr=self.ns.addr())
-        self._controller.init_parameters(ns=self.ns, backend=self.backend)
-        self._logger.init_parameters(
-            log_filename=self.log_filename, save_logfile=self.save_logfile
-        )
-
-    def start_server_mesa(self):
-        """Starts a new AgentNetwork for Mesa and initializes :attr:`_controller`
-
-        Handles the initialisation for :attr:`backend` ``== "mesa"``. Involves
-        spawning two nested objects :attr:`mesa_model` and :attr:`_controller` and
-        calls :meth:`start_mesa_timer`.
-        """
-        self.mesa_model = MesaModel()
-        self._controller = _AgentController(
-            name="AgentController", backend=self.backend
-        )
-        self._controller.init_parameters(
-            backend=self.backend, mesa_model=self.mesa_model
-        )
-        self.start_mesa_timer(self.mesa_update_interval)
-
-    def _set_mode(self, state):
-        """
-        Internal method to set mode of Agent Controller
-        Parameters
-        ----------
-        state: str
-            State of AgentController to set.
-        """
-
-        self._get_controller().set_attr(current_state=state)
-
-    def _get_mode(self):
-        """
-        Returns
-        -------
-        state: str
-            State of Agent Network
-        """
-
-        return self._get_controller().get_attr("current_state")
-
-    def get_mode(self):
-        """
-        Returns
-        -------
-        state: str
-            State of Agent Network
-        """
-
-        return self._get_controller().get_attr("current_state")
-
-    def set_running_state(self, filter_agent=None):
-        """Blanket operation on all agents to set their `current_state` to "Running"
-
-        Users will need to define their own flow of handling each type of
-        `self.current_state` in the `agent_loop`.
-
-        Parameters
-        ----------
-        filter_agent : str
-            (Optional) Filter name of agents to set the states
-
-        """
-
-        self.set_agents_state(filter_agent=filter_agent, state="Running")
-
-    def update_networkx(self):
-        self._get_controller().update_networkx()
-
-    def get_networkx(self):
-        return self._get_controller().get_attr("G")
-
-    def get_nodes_edges(self):
-        G = self.get_networkx()
-        return G.nodes, G.edges(data=True)
-
-    def get_nodes(self):
-        G = self.get_networkx()
-        return G.nodes
-
-    def get_edges(self):
-        G = self.get_networkx()
-        return G.edges
-
-    def set_stop_state(self, filter_agent=None):
-        """
-        Blanket operation on all agents to set their `current_state` attribute to "Stop"
-
-        Users will need to define their own flow of handling each type of
-        `self.current_state` in the `agent_loop`.
-
-        Parameters
-        ----------
-        filter_agent : str
-            (Optional) Filter name of agents to set the states
-
-        """
-
-        self.set_agents_state(filter_agent=filter_agent, state="Stop")
-
-    def set_agents_state(self, filter_agent=None, state="Idle"):
-        """Blanket operation on all agents to set their `current_state` to given state
-
-        Can be used to define different states of operation such as "Running",
-        "Idle, "Stop", etc.. Users will need to define their own flow of handling
-        each type of `self.current_state` in the `agent_loop`.
-
-        Parameters
-        ----------
-        filter_agent : str
-            (Optional) Filter name of agents to set the states
-
-        state : str
-            State of agents to set
-
-        """
-
-        self._set_mode(state)
-        for agent_name in self.agents():
-            if (filter_agent is not None and filter_agent in agent_name) or (
-                filter_agent is None
-            ):
-                agent = self.get_agent(agent_name)
-                try:
-                    agent.set_attr(current_state=state)
-                except Exception as e:
-                    print(e)
-
-        print("SET STATE:  ", state)
-        return 0
-
-    def reset_agents(self):
-        for agent_name in self.agents():
-            agent = self.get_agent(agent_name)
-            agent.reset()
-            agent.set_attr(current_state="Reset")
-        self._set_mode("Reset")
-        return 0
-
-    def remove_agent(self, agent):
-        if type(agent) == str:
-            agent_proxy = self.get_agent(agent)
-        else:
-            agent_proxy = agent
-
-        for input_agent in agent_proxy.get_attr("Inputs"):
-            self.get_agent(input_agent).unbind_output(agent_proxy)
-        for output_agent in agent_proxy.get_attr("Outputs"):
-            agent_proxy.unbind_output(self.get_agent(output_agent))
-
-        agent_proxy.shutdown()
-
-    def bind_agents(self, source, target, channel="default"):
-        """Binds two agents' communication channel in a unidirectional manner
-
-        Any subsequent calls of `source.send_output()` will reach `target` agent's
-        message queue.
-
-        Parameters
-        ----------
-        source : AgentMET4FOF
-            Source agent whose Output channel will be binded to `target`
-
-        target : AgentMET4FOF
-            Target agent whose Input channel will be binded to `source`
-        """
-
-        source.bind_output(target, channel=channel)
-
-        return 0
-
-    def unbind_agents(self, source, target):
-        """Unbinds two agents communication channel in a unidirectional manner
-
-        This is the reverse of `bind_agents()`
-
-        Parameters
-        ----------
-        source : AgentMET4FOF
-            Source agent whose Output channel will be unbinded from `target`
-
-        target : AgentMET4FOF
-            Target agent whose Input channel will be unbinded from `source`
-        """
-
-        source.unbind_output(target)
-        return 0
-
-    def _get_controller(self):
-        """Internal method to access the AgentController relative to the nameserver"""
-        return self._controller
-
-    def _get_logger(self):
-        """Internal method to access the Logger relative to the nameserver"""
-        return self._logger
-
-    def get_agent(self, agent_name):
-        """
-        Returns a particular agent connected to Agent Network.
-
-        Parameters
-        ----------
-        agent_name : str
-            Name of agent to search for in the network
-
-        """
-
-        return self._get_controller().get_agent(agent_name)
-
-    def agents(self, filter_agent: Optional[str] = None) -> List[str]:
-        """Returns all or subset of agents' names connected to agent network
-
-        Parameters
-        ----------
-        filter_agent : str, optional
-            if present, only those names are returned which contain
-            ``filter_agent``'s value
-
-        Returns
-        -------
-        list[str]
-            requested names of agents
-        """
-        all_agent_names = self._get_controller().agents()
-        if filter_agent is not None:
-            filtered_agent_names = [
-                agent_name
-                for agent_name in all_agent_names
-                if filter_agent in agent_name
-            ]
-            return filtered_agent_names
-        return all_agent_names
-
-    def generate_module_name_byType(self, agentType):
-        return self._get_controller().generate_module_name_byType(agentType)
-
-    def add_agent(
-        self,
-        name=" ",
-        agentType=AgentMET4FOF,
-        log_mode=True,
-        buffer_size=1000,
-        ip_addr=None,
-        loop_wait=None,
-        **kwargs,
-    ):
-        """
-        Instantiates a new agent in the network.
-
-        Parameters
-        ----------
-        name str : (Optional) Unique name of agent. here cannot be more than one agent
-            with the same name. Defaults to the agent's class name.
-        agentType AgentMET4FOF : (Optional) Agent class to be instantiated in the
-            network. Defaults to :py:class:`AgentMET4FOF`
-        log_mode bool : (Optional) Determines if messages will be logged to background
-            Logger Agent. Defaults to `True`.
-
-        Returns
-        -------
-        AgentMET4FOF : Newly instantiated agent
-
-        """
-
-        if ip_addr is None:
-            ip_addr = self.ip_addr
-
-        agent = self._get_controller().add_agent(
-            name=name,
-            agentType=agentType,
-            log_mode=log_mode,
-            buffer_size=buffer_size,
-            ip_addr=ip_addr,
-            loop_wait=loop_wait,
-            **kwargs,
-        )
-
-        return agent
-
-    def add_coalition(self, name="Coalition_1", agents=[]):
-        """
-        Instantiates a coalition of agents.
-        """
-        new_coalition = Coalition(name, agents)
-        self._get_controller().add_coalition(new_coalition)
-        return new_coalition
-
-    def add_coalition_agent(self, name="Coalition_1", agents=[]):
-        """
-        Add agents into the coalition
-        """
-        self._get_controller().add_coalition_agent(name, agents)
-
-    def remove_coalition_agent(self, coalition_name, agent_name=""):
-        """
-        Remove agent from coalition
-        """
-        self._get_controller().remove_coalition_agent(coalition_name, agent_name)
-
-    def get_coalition(self, name):
-        """
-        Returns the coalition with the provided name
-        """
-        return self._get_controller().get_coalition(name)
-
-    def del_coalition(self):
-        self._get_controller().del_coalition()
-
-    @property
-    def coalitions(self):
-        return self._get_controller().get_attr("coalitions")
-
-    def get_mesa_model(self):
-        return self.mesa_model
-
-    def shutdown(self):
-        """Shuts down the entire agent network and all agents"""
-
-        # Shutdown the nameserver.
-        # This leaves some process clutter in the process list, but the actual
-        # processes are ended.
-        if self.backend == "osbrain":
-            self._get_controller().get_attr("ns").shutdown()
-        elif self.backend == "mesa":
-            self._get_controller().stop_mesa_timer()
-            self.mesa_model.shutdown()
-
-        # Shutdown the dashboard if present.
-        if self.dashboard_proc is not None:
-            # This calls either the provided method Process.terminate() which
-            # abruptly stops the running multiprocess.Process in case of the osBrain
-            # backend or the self-written method in the class AgentDashboardThread
-            # ensuring the proper termination of the dash.Dash app.
-            self.dashboard_proc.terminate()
-            # Then wait for the termination of the actual thread or at least finish the
-            # execution of the join method in case of the "Mesa" backend. See #163
-            # for the search for a proper solution to this issue.
-            self.dashboard_proc.join(timeout=10)
-        return 0
-
-    def start_mesa_timer(self, update_interval):
-        self._get_controller().start_mesa_timer(update_interval)
-
-    def stop_mesa_timer(self):
-        self._get_controller().stop_mesa_timer()
-
-    def step_mesa_model(self):
-        self._get_controller().step_mesa_model()
-
-
-class Coalition:
-    """
-    A special class for grouping agents.
-
-    It is rendered as a parent group on the dashboard, along with its member agents.
-
-    """
-
-    def __init__(self, name="Coalition", agents=[]):
-        self.agents = agents
-        self.name = name
-
-    def agent_names(self):
-        return [agent.get_attr("name") for agent in self.agents]
-
-    def add_agent(self, agent):
-        self.agents.append(agent)
-
-    def remove_agent(self, agent):
-        if isinstance(agent, str):
-            self.agents = [agent_i for agent_i in self.agents if agent_i.name != agent]
-        elif isinstance(agent, AgentMET4FOF):
-            self.agents = [
-                agent_i for agent_i in self.agents if agent_i.name != agent.name
-            ]
-
-
 class DataStreamAgent(AgentMET4FOF):
     """Able to simulate generation of datastream by loading a given DataStreamMET4FOF
 
@@ -2276,390 +1065,3 @@ class MonitorAgent(AgentMET4FOF):
         self.plots = {}
 
 
-class _Logger(AgentMET4FOF):
-    """An internal logger agent which is instantiated with each AgentNetwork
-
-    It collects all the logs which are sent to it, and print them and optionally
-    save them into a csv log file. Since the user is not expected to directly access
-    the logger agent, its initialisation option and interface are provided via the
-    AgentNetwork object.
-
-    When log_info of any agent is called, the agent will send the data to the logger
-    agent.
-    """
-
-    def init_parameters(self, log_filename="log_file.csv", save_logfile=True):
-        self.current_log_handlers = {"INFO": self.log_handler}
-        self.bind("SUB", "sub", {"INFO": self.log_handler})
-        self.log_filename = log_filename
-        self.save_logfile = save_logfile
-        if self.save_logfile:
-            try:
-                # writes a new file
-                self.writeFile = open(self.log_filename, "w", newline="")
-                writer = csv.writer(self.writeFile)
-                writer.writerow(["Time", "Name", "Topic", "Data"])
-                # set to append mode
-                self.writeFile = open(self.log_filename, "a", newline="")
-            except:
-                raise Exception
-        self.save_cycles = 0
-
-    @property
-    def subscribed_topics(self):
-        return list(self.current_log_handlers.keys())
-
-    def bind_log_handler(self, log_handler_functions):
-        for topic in self.subscribed_topics:
-            self.unsubscribe("sub", topic)
-        self.current_log_handlers.update(log_handler_functions)
-        self.subscribe("sub", self.current_log_handlers)
-
-    def log_handler(self, message, topic):
-        sys.stdout.write(message + "\n")
-        sys.stdout.flush()
-        self.save_log_info(str(message))
-
-    def save_log_info(self, log_msg):
-        re_sq = r"\[(.*?)\]"
-        re_rd = r"\((.*?)\)"
-
-        date = re.findall(re_sq, log_msg)[0]
-        date = "[" + date + "]"
-
-        agent_name = re.findall(re_rd, log_msg)[0]
-
-        contents = log_msg.split(":")
-        if len(contents) > 4:
-            topic = contents[3]
-            data = str(contents[4:])
-        else:
-            topic = contents[3]
-            data = " "
-
-        if self.save_logfile:
-            try:
-                # append new row
-                writer = csv.writer(self.writeFile)
-                writer.writerow([str(date), agent_name, topic, data])
-
-                if self.save_cycles % 15 == 0:
-                    self.writeFile.close()
-                    self.writeFile = open(self.log_filename, "a", newline="")
-                self.save_cycles += 1
-            except:
-                raise Exception
-
-
-class MetrologicalAgent(AgentMET4FOF):
-    # dict like {
-    #     <from>: {
-    #         "buffer": TimeSeriesBuffer(maxlen=buffer_size),
-    #         "metadata": MetaData(**kwargs).metadata,
-    #     }
-    _input_data: Dict[str, Dict[str, Union[TimeSeriesBuffer, Dict]]]
-    """Input dictionary of all incoming data including metadata::
-
-        dict like {
-            <from>: {
-                "buffer": TimeSeriesBuffer(maxlen=buffer_size),
-                "metadata": MetaData(**kwargs).metadata,
-            }
-    """
-    _input_data_maxlen: int
-
-    # dict like {
-    #     <channel> : {
-    #         "buffer" : TimeSeriesBuffer(maxlen=buffer_size),
-    #         "metadata" : MetaData(**kwargs)
-    #     }
-    _output_data: Dict[str, Dict[str, Union[TimeSeriesBuffer, MetaData]]]
-    """Output dictionary of all outgoing data including metadata::
-
-        dict like {
-            <from>: {
-                "buffer": TimeSeriesBuffer(maxlen=buffer_size),
-                "metadata": MetaData(**kwargs).metadata,
-            }
-    """
-    _output_data_maxlen: int
-
-    def init_parameters(self, input_data_maxlen=25, output_data_maxlen=25):
-        super(MetrologicalAgent, self).init_parameters()
-        self._input_data = {}
-        self._input_data_maxlen = input_data_maxlen
-        self._output_data = {}
-        self._output_data_maxlen = output_data_maxlen
-
-    def on_received_message(self, message):
-        channel = message["channel"]
-        sender = message["from"]
-
-        if channel == "default":
-            data = message["data"]
-            metadata = None
-            if "metadata" in message.keys():
-                metadata = message["metadata"]
-
-            self._set_input_data(sender, data, metadata)
-
-    def _set_input_data(self, sender, data=None, metadata=None):
-        # create storage for new senders
-        if sender not in self._input_data.keys():
-            self._input_data[sender] = {
-                "metadata": metadata,
-                "buffer": TimeSeriesBuffer(maxlen=self._input_data_maxlen),
-            }
-
-        if metadata is not None:
-            # update received metadata
-            self._input_data[sender]["metadata"] = metadata
-
-        if data is not None:
-            # append received data
-            self._input_data[sender]["buffer"].add(data=data)
-
-    def set_output_data(self, channel, data=None, metadata=None):
-        # create storage for new output channels
-        if channel not in self._output_data.keys():
-            self._output_data[channel] = {
-                "metadata": metadata,
-                "buffer": TimeSeriesBuffer(maxlen=self._output_data_maxlen),
-            }
-
-        if metadata is not None:
-            # update received metadata
-            self._output_data[channel]["metadata"] = metadata
-
-        if data is not None:
-            # append received data
-            self._output_data[channel]["buffer"].add(data=data)
-
-    def agent_loop(self):
-        if self.current_state == "Running":
-
-            for channel, channel_dict in self._output_data.items():
-                # short names
-                metadata = channel_dict["metadata"]
-                buffer = channel_dict["buffer"]
-
-                # if there is something in the buffer, send it all
-                buffer_len = len(buffer)
-                if buffer_len > 0:
-                    data = buffer.pop(n_samples=buffer_len)
-
-                    # send data+metadata
-                    self.send_output([data, metadata], channel=channel)
-
-    def pack_data(self, data, channel="default"):
-
-        # include metadata in the packed data
-        packed_data = {
-            "from": self.name,
-            "data": data[0],
-            "metadata": data[1],
-            "senderType": type(self).__name__,
-            "channel": channel,
-        }
-
-        return packed_data
-
-
-class MetrologicalMonitorAgent(MetrologicalAgent):
-    def init_parameters(self, *args, **kwargs):
-        super(MetrologicalMonitorAgent, self).init_parameters(*args, **kwargs)
-        # create alias/dummies to match dashboard expectations
-        self.memory = self._input_data
-        self.plot_filter = []
-        self.plots = {}
-        self.custom_plot_parameters = {}
-
-    def on_received_message(self, message):
-        """
-        Handles incoming data from 'default' and 'plot' channels.
-
-        Stores 'default' data into `self.memory` and 'plot' data into `self.plots`
-
-        Parameters
-        ----------
-        message : dict
-            Acceptable channel values are 'default' or 'plot'
-        """
-        if message["channel"] == "default":
-            if self.plot_filter != []:
-                message["data"] = {
-                    key: message["data"][key] for key in self.plot_filter
-                }
-                message["metadata"] = {
-                    key: message["metadata"][key] for key in self.plot_filter
-                }
-            self.buffer_store(
-                agent_from=message["from"],
-                data={"data": message["data"], "metadata": message["metadata"]},
-            )
-        elif message["channel"] == "plot":
-            self.update_plot_memory(message)
-        return 0
-
-    def update_plot_memory(self, message):
-        """
-        Updates plot figures stored in `self.plots` with the received message
-
-        Parameters
-        ----------
-        message : dict
-            Standard message format specified by AgentMET4FOF class
-            Message['data'] needs to be base64 image string and can be nested in dictionary for multiple plots
-            Only the latest plot will be shown kept and does not keep a history of the plots.
-        """
-
-        if type(message["data"]) != dict or message["from"] not in self.plots.keys():
-            self.plots[message["from"]] = message["data"]
-        elif type(message["data"]) == dict:
-            for key in message["data"].keys():
-                self.plots[message["from"]].update({key: message["data"][key]})
-        self.log_info("PLOTS: " + str(self.plots))
-
-    def reset(self):
-        super(MetrologicalMonitorAgent, self).reset()
-        del self.plots
-        self.plots = {}
-
-    def custom_plot_function(self, data, sender_agent, **kwargs):
-        # TODO: cannot set the label of the xaxis within this method
-        # data display
-        if "data" in data.keys():
-            if len(data["data"]):
-                # values = data["buffer"].show(n_samples=-1)  # -1 --> all
-                values = data["data"]
-                t = values[:, 0]
-                ut = values[:, 1]
-                v = values[:, 2]
-                uv = values[:, 3]
-
-                # use description
-                desc = data["metadata"][0]
-                t_name, t_unit = desc.time.values()
-                v_name, v_unit = desc.get_quantity().values()
-
-                x_label = f"{t_name} [{t_unit}]"
-                y_label = f"{v_name} [{v_unit}]"
-
-                trace = go.Scatter(
-                    x=t,
-                    y=v,
-                    error_x=dict(type="data", array=ut, visible=True),
-                    error_y=dict(type="data", array=uv, visible=True),
-                    mode="lines",
-                    name=f"{y_label} ({sender_agent})",
-                )
-            else:
-                trace = go.Scatter()
-        else:
-            trace = go.Scatter()
-        return trace
-
-
-class MetrologicalAgentBuffer(AgentBuffer):
-    """Buffer class which is instantiated in every metrological agent to store data
-
-    This buffer is necessary to handle multiple inputs coming from agents.
-
-    We can access the buffer like a dict with exposed functions such as .values(),
-    .keys() and .items(). The actual dict object is stored in the attribute
-    :attr:`buffer <agentMET4FOF.agents.AgentBuffer.buffer>`. The list in
-    :attr:`supported_datatypes <agentMET4FOF.agents.AgentBuffer.supported_datatypes>`
-    contains one more element
-    for metrological agents, namely :class:`TimeSeriesBuffer
-    <time-series-buffer:time_series_buffer.buffer.TimeSeriesBuffer>`.
-    """
-
-    def __init__(self, buffer_size: int = 1000):
-        """Initialise a new agent buffer object
-
-        Parameters
-        ----------
-        buffer_size: int
-            Length of buffer allowed.
-        """
-        super(MetrologicalAgentBuffer, self).__init__(buffer_size)
-        self.supported_datatypes.append(TimeSeriesBuffer)
-
-    def convert_single_to_tsbuffer(self, single_data: Union[List, Tuple, np.ndarray]):
-        """Convert common data in agentMET4FOF to :class:`TimeSeriesBuffer
-        <time-series-buffer:time_series_buffer.buffer.TimeSeriesBuffer>`
-
-        Parameters
-        ----------
-        single_data : iterable of iterables (list, tuple, np.ndarrray) with shape (N, M)
-
-            * M==2 (pairs): assumed to be like (time, value)
-            * M==3 (triple): assumed to be like (time, value, value_unc)
-            * M==4 (4-tuple): assumed to be like (time, time_unc, value, value_unc)
-
-        Returns
-        -------
-        TimeSeriesBuffer
-            the new :class:`TimeSeriesBuffer
-            <time-series-buffer:time_series_buffer.buffer.TimeSeriesBuffer>` object
-
-        """
-        ts = TimeSeriesBuffer(maxlen=self.buffer_size)
-        ts.add(single_data)
-        return ts
-
-    def update(
-        self,
-        agent_from: str,
-        data: Union[Dict, List, Tuple, np.ndarray],
-    ) -> TimeSeriesBuffer:
-        """Overrides data in the buffer dict keyed by `agent_from` with value `data`
-
-        Parameters
-        ----------
-        agent_from : str
-            Name of agent sender
-        data : dict or iterable of iterables (list, tuple, np.ndarray) with shape (N, M
-            the data to be stored in the metrological buffer
-
-        Returns
-        -------
-        TimeSeriesBuffer
-            the updated :class:`TimeSeriesBuffer
-            <time-series-buffer:time_series_buffer.buffer.TimeSeriesBuffer>` object
-        """
-        # handle if data type nested in dict
-        if isinstance(data, dict):
-            # check for each value datatype
-            for key, value in data.items():
-                data[key] = self.convert_single_to_tsbuffer(value)
-        else:
-            data = self.convert_single_to_tsbuffer(data)
-            self.buffer.update({agent_from: data})
-        return self.buffer
-
-    def _concatenate(
-        self,
-        iterable: TimeSeriesBuffer,
-        data: Union[np.ndarray, list, pd.DataFrame],
-        concat_axis: int = 0,
-    ) -> TimeSeriesBuffer:
-        """Concatenate the given ``TimeSeriesBuffer`` with ``data``
-
-        Add ``data`` to the :class:`TimeSeriesBuffer
-        <time-series-buffer:time_series_buffer.buffer.TimeSeriesBuffer>` object.
-
-        Parameters
-        ----------
-        iterable : TimeSeriesBuffer
-            The current buffer to be concatenated with.
-        data : np.ndarray, DataFrame, list
-            New incoming data
-
-        Returns
-        -------
-        TimeSeriesBuffer
-            the original buffer with the data appended
-        """
-        iterable.add(data)
-        return iterable
