+++ agentMET4FOF/metrological_streams/__init__.py
@@ -4,14 +4,14 @@ Up until version 0.8.1 we had several modules instead of the packages `agents` a
 `streams`. We have included this package to not break the previously needed import
 statements.
 """
-from ..streams.base_streams import MetrologicalDataStreamMET4FOF
+from ..streams.metrological_base_streams import MetrologicalDataStreamMET4FOF
 from ..streams.signal_streams import (
     MetrologicalMultiWaveGenerator,
     MetrologicalSineGenerator,
 )
 
 __all__ = [
-    "MetrologicalDataStreamMET4FOF",
     "MetrologicalMultiWaveGenerator",
     "MetrologicalSineGenerator",
+    "MetrologicalDataStreamMET4FOF",
 ]
