+++ tests/conftest.py
@@ -3,7 +3,8 @@
 import numpy as np
 import pytest
 
-from agentMET4FOF.agents.base_agents import AgentNetwork
+from agentMET4FOF.agents import AgentMET4FOF
+from agentMET4FOF.agents.network import AgentNetwork
 
 # Set time to wait for before agents should have done their jobs in networks.
 test_timeout = 20
