+++ tests/test_addremove_metrological_agents.py
@@ -1,6 +1,9 @@
 import pytest
 
-from agentMET4FOF.agents.base_agents import MetrologicalAgent, MetrologicalMonitorAgent
+from agentMET4FOF.agents.metrological_base_agents import (
+    MetrologicalAgent,
+    MetrologicalMonitorAgent,
+)
 from tests.conftest import test_timeout
 
 agent_type_permutations = [
